View collections by proxy
Proof of concept.  Attempts to display `/collections` page when a
`/_collections` page is visited. If no such `Thing` exists, an HTTP
error is raised and nothing is cached.  Edit and history links on cached
page point to the `/collections` object, so there's no need to handle
every mode in the proxy handler.